% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HNSCImmunophenotype.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{HNSCImmunophenotype data}
\format{
list:
\describe{
\item{geo.neg.gene}{the expression matrix of the GSE65858 HPV-}
\item{geo.neg.type}{the immunophenotype of the GSE65858 HPV-}
\item{geo.pos.gene}{the expression matrix of the GSE65858 HPV+}
\item{geo.pos.type}{the immunophenotype of the GSE65858 HPV+}
\item{tcga.neg.gene}{the expression matrix of the TCGA-HNSC HPV-}
\item{tcga.neg.type}{the immunophenotype of the TCGA-HNSC HPV-}
\item{tcga.pos.gene}{the expression matrix of the TCGA-HNSC HPV+}
\item{tcga.pos.type}{the expression matrix of the TCGA-HNSC HPV+}
}
}
\usage{
dataset
}
\description{
HNSCImmunophenotype data
}
\keyword{datasets}
