% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HNSCImmunophenotype.R
\docType{package}
\name{HNSCImmunophenotype}
\alias{HNSCImmunophenotype}
\title{HNSCImmunophenotype:A package for calculating HNSC immunophenotype.}
\usage{
HNSCImmunophenotype(exp, hpv, ntree = 10000, set.seed = 123)
}
\arguments{
\item{exp}{the gene expression data set. A matrix with row names as symbols and columns as samples.}

\item{hpv}{the HPV infection information,There are "positive" and "negative" can been choiced.}

\item{ntree}{The number of decision trees contained in the random forest,defaults = 10000.}

\item{set.seed}{Set random seeds so that the results can be repeated,defaults = 1234.}
}
\value{
the immunophenotype.
}
\description{
\code{HNSCImmunophenotype} Returns the HNSC immunophenotype.
}
